lambda/drift/requirements.txt
numpy
scipy
pandas
mysql-connector-python
boto3
aws-lambda-powertools[parser]
